feat(api-capabilities): introduce HandlesTask trait for task-related methods
